\name{mabc}
\alias{mabc}
\docType{package}
\title{
Maciel's Artificial Bee Colony algorithm implementation
}

\description{
%%\packageDescription{mabc}
Implementation of the Artificial Bee Colony algorithm.
}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{mabc}
\packageIndices{mabc}
~~ An overview of how to use the package, including the most important functions ~~

}
\author{
\packageAuthor{mabc}

Maintainer: \packageMaintainer{mabc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
