% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.fit.parallel.R
\name{mice.fit.parallel}
\alias{mice.fit.parallel}
\title{MICE with weights in parallel (multiple cores of 1 node)}
\usage{
mice.fit.parallel(data, mice.model = mice.fit.wrapper, m = 1,
  n_cores = 1, n_experiments, method = "norm",
  defaultMethod = "norm", predictorMatrix, maxit = 5,
  printFlag = TRUE, seed.init = 0)
}
\arguments{
\item{data}{training dataframe}

\item{mice.model}{wrapper function for mice}

\item{m}{number of imputations}

\item{n_cores}{number of cores available for parallel running of mice}

\item{n_experiments}{number of experiments (typically a multiple of n.experiments)}

\item{method}{string (vector)}

\item{defaultMethod}{c("pmm", "logreg", "polyreg", "polr")}

\item{predictorMatrix}{predictorMatrix}

\item{maxit}{How many cycles through chained equations}

\item{printFlag}{Do you want stuff printed}

\item{seed.init}{seed value to make it reproducible}
}
\value{
A list with imputations and their attributes, including their inverse probability weights
}
\description{
Produces a list with imputations with inverse probability weights.
}
