% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.fit.R
\name{mice.fit}
\alias{mice.fit}
\title{MICE with weights}
\usage{
mice.fit(data, m = 5, method = "norm", predictorMatrix, where = NULL,
  blocks, visitSequence = NULL, formulas, blots = NULL, post = NULL,
  defaultMethod = "norm", maxit = 5, printFlag = TRUE, seed = NA,
  data.init = NULL, ...)
}
\arguments{
\item{data}{training dataframe}

\item{m}{number of imputations}

\item{method}{string (vector)}

\item{predictorMatrix}{predictorMatrix}

\item{blocks}{blocks}

\item{visitSequence}{defaults is 1 to ncol(data)}

\item{formulas}{formulas}

\item{blots}{blots}

\item{post}{post}

\item{defaultMethod}{c("pmm", "logreg", "polyreg", "polr")}

\item{maxit}{How many cycles through chained equations}

\item{printFlag}{Do you want stuff printed}

\item{seed}{seed value to make it reproducible}

\item{data.init}{Same}
}
\value{
A list with imputations and their attributes, including their inverse probability weights
}
\description{
Produces a list with imputations with inverse probability weights.
}
