% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.fit.wrapper.R
\name{mice.fit.wrapper}
\alias{mice.fit.wrapper}
\title{mice.fit wrapper}
\usage{
mice.fit.wrapper(mice.fit.input.list)
}
\arguments{
\item{mice.fit.input.list}{list of stuff that mice.fit needs}
}
\value{
A list with imputations and their attributes, including their inverse probability weights
}
\description{
Wraps the mice.fit function for parallel execution.
}
