% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.parallel.R
\name{mice.parallel}
\alias{mice.parallel}
\title{MICE with weights in parallel (multiple cores of 1 node)}
\usage{
mice.parallel(data, m = 1, method = "norm", predictorMatrix,
  maxit = 5, printFlag = FALSE, n_cores = 1, n_experiments)
}
\arguments{
\item{data}{training dataframe}

\item{m}{number of imputations (1)}

\item{method}{string (vector)}

\item{predictorMatrix}{predictorMatrix}

\item{maxit}{How many cycles through chained equations}

\item{printFlag}{Do you want stuff printed}

\item{n_cores}{number of cores available for parallel running of mice}

\item{n_experiments}{number of experiments (typically a multiple of n.experiments)}
}
\value{
A list with imputations and their attributes
}
\description{
Produces a list with imputations with inverse probability weights.
}
