% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.mpi.run.R
\name{model.mpi.run}
\alias{model.mpi.run}
\title{Distribute model runs over multiple nodes}
\usage{
model.mpi.run(model, actual.input.matrix, seed_count = 0, slaves = 3)
}
\arguments{
\item{model}{Wrapper function for the model}

\item{actual.input.matrix}{Matrix with parameter combinations to be run}

\item{seed_count}{Origin of random number seed}

\item{slaves}{Number of slave workers available for parallel running of the model}
}
\value{
a matrix of model features and the seed of the random number
  generator
}
\description{
Allows faster execution of each wave of simulations
}
